@@ -1,6 +1,6 @@
 {
   "name": "react-formio",
-  "version": "4.2.3",
+  "version": "5.0.0-alpha.1",
   "description": "React renderer for form.io forms.",
   "main": "lib/index.js",
   "scripts": {
@@ -31,17 +31,15 @@
     "babel-polyfill": "^6.26.0",
     "core-js": "^3.1.4",
     "eventemitter2": "^5.0.1",
-    "formiojs": "https://github.com/Casecommons/formio.js.git",
     "lodash": "^4.17.15",
-    "prop-types": "^15.7.2",
-    "recompose": "^0.30.0"
+    "prop-types": "^15.7.2"
   },
   "devDependencies": {
     "ajv-pack": "^0.2.5",
     "babel-cli": "^6.24.1",
     "babel-core": "^6.26.3",
     "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.4",
+    "babel-loader": "^8.0.6",
     "babel-preset-es2015": "^6.14.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-2": "^6.13.0",
@@ -58,16 +56,17 @@
     "eslint-plugin-jsx-a11y": "^2.0.1",
     "eslint-plugin-mocha": "^5.2.0",
     "eslint-plugin-react": "^6.0.0",
+    "formiojs": "https://github.com/Casecommons/formio.js.git",
     "jsdom": "^9.4.2",
-    "karma": "^3.0.0",
+    "karma": "^4.4.1",
     "karma-chrome-launcher": "^2.2.0",
     "karma-mocha": "^1.3.0",
     "karma-mocha-reporter": "^2.2.5",
     "karma-webpack": "^3.0.5",
     "mocha": "^5.2.0",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-test-renderer": "^16.5.2",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-test-renderer": "^16.13.1",
     "react-transform-catch-errors": "^1.0.0",
     "sinon": "^7.0.0",
     "webpack": "^2.7.0"
@@ -76,6 +75,7 @@
     "lib"
   ],
   "peerDependencies": {
+    "formiojs": "https://github.com/Casecommons/formio.js.git",
     "react": "^15.3.0 || ^16.0.0",
     "react-dom": "^15.3.0 || ^16.0.0"
   }
