@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React, {useEffect} from 'react';
 import PropTypes from 'prop-types';
 import AllComponents from 'formiojs/lib/components';
 import Components from 'formiojs/lib/components/Components';
@@ -6,89 +6,79 @@ import FormioFormBuilder from 'formiojs/lib/FormBuilder';
 
 Components.setComponents(AllComponents);
 
-export default class FormBuilder extends Component {
-  static defaultProps = {
-    options: {},
-    Builder: FormioFormBuilder
-  };
+const FormBuilder = (props) => {
+  let builder;
+  let builderReady;
+  let element;
 
-  static propTypes = {
-    form: PropTypes.object,
-    options: PropTypes.object,
-    onSaveComponent: PropTypes.func,
-    onUpdateComponent: PropTypes.func,
-    onDeleteComponent: PropTypes.func,
-    onCancelComponent: PropTypes.func,
-    onEditComponent: PropTypes.func,
-    Builder: PropTypes.any
-  };
-
-  componentDidMount = () => {
-    this.initializeBuilder(this.props);
-  };
-
-  componentWillUnmount = () => {
-    if (this.builder !== undefined) {
-      this.builder.instance.destroy(true);
+  const emit = (funcName) => (...args) => {
+    if (props.hasOwnProperty(funcName) && typeof (props[funcName]) === 'function') {
+      props[funcName](...args);
     }
   };
 
-  initializeBuilder = (props) => {
-    const options = Object.assign({}, props.options);
-    const form = Object.assign({}, props.form);
-    const Builder = props.Builder;
-
-    if (this.builder !== undefined) {
-      this.builder.instance.destroy(true);
+  const onChange = () => {
+    const {onChange} = props;
+    if (onChange && typeof onChange === 'function') {
+      onChange(builder.instance.form);
     }
+  };
 
-    this.builder = new Builder(this.element.firstChild, form, options);
-    this.builderReady = this.builder.ready;
+  const builderEvents = [
+    {name: 'saveComponent', action: emit('onSaveComponent')},
+    {name: 'updateComponent', action: emit('onUpdateComponent')},
+    {name: 'removeComponent', action: emit('onDeleteComponent')},
+    {name: 'cancelComponent', action: emit('onUpdateComponent')},
+    {name: 'editComponent', action: emit('onEditComponent')},
+    {name: 'addComponent', action: onChange},
+    {name: 'saveComponent', action: onChange},
+    {name: 'updateComponent', action: onChange},
+    {name: 'removeComponent', action: onChange},
+    {name: 'deleteComponent', action: onChange},
+    {name: 'pdfUploaded', action: onChange},
+  ];
 
-    this.builderReady.then(() => {
-      this.builder.instance.on('saveComponent', this.emit('onSaveComponent'));
-      this.builder.instance.on('updateComponent', this.emit('onUpdateComponent'));
-      this.builder.instance.on('removeComponent', this.emit('onDeleteComponent'));
-      this.builder.instance.on('cancelComponent', this.emit('onCancelComponent'));
-      this.builder.instance.on('editComponent', this.emit('onEditComponent'));
-      this.builder.instance.on('addComponent', this.onChange);
-      this.builder.instance.on('saveComponent', this.onChange);
-      this.builder.instance.on('updateComponent', this.onChange);
-      this.builder.instance.on('removeComponent', this.onChange);
-      this.builder.instance.on('deleteComponent', this.onChange);
-      this.builder.instance.on('pdfUploaded', this.onChange);
+  const initializeBuilder = (builderProps) => {
+    let {options, form} = builderProps;
+    const {Builder} = builderProps;
+    options = Object.assign({}, options);
+    form = Object.assign({}, form);
+
+    builder = new Builder(element.firstChild, form, options);
+    builderReady = builder.ready;
+
+    builderReady.then(() => {
+      onChange();
+      builderEvents.forEach(({name, action}) => builder.instance.on(name, action));
     });
   };
 
-  componentWillReceiveProps = (nextProps) => {
-    const {options, form} = this.props;
+  useEffect(() => {
+    initializeBuilder(props);
+    return () => builder ? builder.instance.destroy(true) : null;
+  }, [props.form.display, props.form.components, props.options]);
 
-    if (
-      (form.display !== nextProps.form.display)
-      || (options !== nextProps.options)
-      || (form.components !== nextProps.form.components)
-    ) {
-      this.initializeBuilder(nextProps);
-    }
-  };
-
-  render = () => {
-    return <div ref={element => this.element = element}>
+  return (
+    <div ref={el => element = el}>
       <div></div>
-    </div>;
-  };
+    </div>
+  );
+};
 
-  onChange = () => {
-    if (this.props.hasOwnProperty('onChange') && typeof this.props.onChange === 'function') {
-      this.props.onChange(this.builder.instance.form);
-    }
-  };
+FormBuilder.defaultProps = {
+  options: {},
+  Builder: FormioFormBuilder
+};
 
-  emit = (funcName) => {
-    return (...args) => {
-      if (this.props.hasOwnProperty(funcName) && typeof (this.props[funcName]) === 'function') {
-        this.props[funcName](...args);
-      }
-    };
-  };
-}
+FormBuilder.propTypes = {
+  form: PropTypes.object,
+  options: PropTypes.object,
+  onSaveComponent: PropTypes.func,
+  onUpdateComponent: PropTypes.func,
+  onDeleteComponent: PropTypes.func,
+  onCancelComponent: PropTypes.func,
+  onEditComponent: PropTypes.func,
+  Builder: PropTypes.any
+};
+
+export default FormBuilder;
