@@ -4,8 +4,6 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
 var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
@@ -32,43 +30,108 @@ var _Form2 = _interopRequireDefault(_Form);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
-
 _Components2.default.setComponents(_components2.default);
 
-var Form = function (_Component) {
-  _inherits(Form, _Component);
 
-  function Form() {
-    var _ref;
+var Form = function Form(props) {
+  var instance = void 0;
+  var createPromise = void 0;
+  var element = void 0;
+  var formio = void 0;
 
-    var _temp, _this, _ret;
+  (0, _react.useEffect)(function () {
+    return function () {
+      return formio ? formio.destroy(true) : null;
+    };
+  }, [formio]);
 
-    _classCallCheck(this, Form);
+  var createWebformInstance = function createWebformInstance(srcOrForm) {
+    var _props$options = props.options,
+        options = _props$options === undefined ? {} : _props$options,
+        formioform = props.formioform;
 
-    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
-      args[_key] = arguments[_key];
+    instance = new (formioform || _Form2.default)(element, srcOrForm, options);
+    createPromise = instance.ready.then(function (formioInstance) {
+      formio = formioInstance;
+    });
+
+    return createPromise;
+  };
+
+  var onAnyEvent = function onAnyEvent(event) {
+    for (var _len = arguments.length, args = Array(_len > 1 ? _len - 1 : 0), _key = 1; _key < _len; _key++) {
+      args[_key - 1] = arguments[_key];
     }
 
-    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_ref = Form.__proto__ || Object.getPrototypeOf(Form)).call.apply(_ref, [this].concat(args))), _this), _initialiseProps.call(_this), _temp), _possibleConstructorReturn(_this, _ret);
-  }
+    if (event.startsWith('formio.')) {
+      var funcName = 'on' + event.charAt(7).toUpperCase() + event.slice(8);
+      if (props.hasOwnProperty(funcName) && typeof props[funcName] === 'function') {
+        props[funcName].apply(props, args);
+      }
+    }
+  };
 
-  _createClass(Form, null, [{
-    key: 'getDefaultEmitter',
-    value: function getDefaultEmitter() {
-      return new _eventemitter2.default({
-        wildcard: false,
-        maxListeners: 0
+  var initializeFormio = function initializeFormio() {
+    var submission = props.submission;
+
+    if (createPromise) {
+      instance.onAny(onAnyEvent);
+      createPromise.then(function () {
+        if (submission) {
+          formio.submission = submission;
+        }
       });
     }
-  }]);
+  };
 
-  return Form;
-}(_react.Component);
+  (0, _react.useEffect)(function () {
+    var src = props.src;
+
+    if (src) {
+      createWebformInstance(src).then(function () {
+        formio.src = src;
+      });
+      initializeFormio();
+    }
+  }, [props.src]);
+
+  (0, _react.useEffect)(function () {
+    var form = props.form,
+        url = props.url;
+
+    if (form) {
+      createWebformInstance(form).then(function () {
+        formio.form = form;
+        if (url) {
+          formio.url = url;
+        }
+        return formio;
+      });
+      initializeFormio();
+    }
+  }, [props.form]);
+
+  (0, _react.useEffect)(function () {
+    var _props$options2 = props.options,
+        options = _props$options2 === undefined ? {} : _props$options2;
+
+    if (!options.events) {
+      options.events = Form.getDefaultEmitter();
+    }
+  }, [props.options]);
+
+  (0, _react.useEffect)(function () {
+    var submission = props.submission;
+
+    if (formio && submission) {
+      formio.submission = submission;
+    }
+  }, [props.submission]);
+
+  return _react2.default.createElement('div', { ref: function ref(el) {
+      return element = el;
+    } });
+};
 
 Form.propTypes = {
   src: _propTypes2.default.string,
@@ -101,115 +164,11 @@ Form.propTypes = {
   formioform: _propTypes2.default.any
 };
 
-var _initialiseProps = function _initialiseProps() {
-  var _this2 = this;
-
-  this.componentDidMount = function () {
-    var _props = _this2.props,
-        _props$options = _props.options,
-        options = _props$options === undefined ? {} : _props$options,
-        src = _props.src,
-        url = _props.url,
-        form = _props.form;
-
-
-    if (!options.events) {
-      options.events = Form.getDefaultEmitter();
-    }
-
-    if (src) {
-      _this2.instance = new (_this2.props.formioform || _Form2.default)(_this2.element, src, options);
-      _this2.createPromise = _this2.instance.ready.then(function (formio) {
-        _this2.formio = formio;
-        _this2.formio.src = src;
-      });
-    }
-    if (form) {
-      _this2.instance = new (_this2.props.formioform || _Form2.default)(_this2.element, form, options);
-      _this2.createPromise = _this2.instance.ready.then(function (formio) {
-        _this2.formio = formio;
-        _this2.formio.form = form;
-        if (url) {
-          _this2.formio.url = url;
-        }
-
-        return _this2.formio;
-      });
-    }
-
-    _this2.initializeFormio();
-  };
-
-  this.componentWillUnmount = function () {
-    if (_this2.formio !== undefined) {
-      _this2.formio.destroy(true);
-    }
-  };
-
-  this.initializeFormio = function () {
-    if (_this2.createPromise) {
-      _this2.instance.onAny(function (event) {
-        for (var _len2 = arguments.length, args = Array(_len2 > 1 ? _len2 - 1 : 0), _key2 = 1; _key2 < _len2; _key2++) {
-          args[_key2 - 1] = arguments[_key2];
-        }
-
-        if (event.startsWith('formio.')) {
-          var funcName = 'on' + event.charAt(7).toUpperCase() + event.slice(8);
-          if (_this2.props.hasOwnProperty(funcName) && typeof _this2.props[funcName] === 'function') {
-            var _props2;
-
-            (_props2 = _this2.props)[funcName].apply(_props2, args);
-          }
-        }
-      });
-      _this2.createPromise.then(function () {
-        if (_this2.props.submission) {
-          _this2.formio.submission = _this2.props.submission;
-        }
-      });
-    }
-  };
-
-  this.componentWillReceiveProps = function (nextProps) {
-    var _props3 = _this2.props,
-        _props3$options = _props3.options,
-        options = _props3$options === undefined ? {} : _props3$options,
-        src = _props3.src,
-        form = _props3.form,
-        submission = _props3.submission;
-
-
-    if (!options.events) {
-      options.events = Form.getDefaultEmitter();
-    }
-
-    if (src !== nextProps.src) {
-      _this2.instance = new (_this2.props.formioform || _Form2.default)(_this2.element, nextProps.src, options);
-      _this2.createPromise = _this2.instance.ready.then(function (formio) {
-        _this2.formio = formio;
-        _this2.formio.src = nextProps.src;
-      });
-      _this2.initializeFormio();
-    }
-    if (form !== nextProps.form) {
-      _this2.instance = new (_this2.props.formioform || _Form2.default)(_this2.element, nextProps.form, options);
-      _this2.createPromise = _this2.instance.ready.then(function (formio) {
-        _this2.formio = formio;
-        _this2.formio.form = nextProps.form;
-      });
-      _this2.initializeFormio();
-    }
-
-    if (submission !== nextProps.submission && _this2.formio) {
-      _this2.formio.submission = nextProps.submission;
-    }
-  };
-
-  this.render = function () {
-    return _react2.default.createElement('div', { ref: function ref(element) {
-        return _this2.element = element;
-      } });
-  };
+Form.getDefaultEmitter = function () {
+  return new _eventemitter2.default({
+    wildcard: false,
+    maxListeners: 0
+  });
 };
 
 exports.default = Form;
