@@ -26,38 +26,73 @@ var _FormBuilder2 = _interopRequireDefault(_FormBuilder);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
-
 _Components2.default.setComponents(_components2.default);
 
-var FormBuilder = function (_Component) {
-  _inherits(FormBuilder, _Component);
+var FormBuilder = function FormBuilder(props) {
+  var builder = void 0;
+  var builderReady = void 0;
+  var element = void 0;
 
-  function FormBuilder() {
-    var _ref;
+  var emit = function emit(funcName) {
+    return function () {
+      if (props.hasOwnProperty(funcName) && typeof props[funcName] === 'function') {
+        props[funcName].apply(props, arguments);
+      }
+    };
+  };
 
-    var _temp, _this, _ret;
+  var onChange = function onChange() {
+    var onChange = props.onChange;
 
-    _classCallCheck(this, FormBuilder);
-
-    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
-      args[_key] = arguments[_key];
+    if (onChange && typeof onChange === 'function') {
+      onChange(builder.instance.form);
     }
+  };
 
-    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_ref = FormBuilder.__proto__ || Object.getPrototypeOf(FormBuilder)).call.apply(_ref, [this].concat(args))), _this), _initialiseProps.call(_this), _temp), _possibleConstructorReturn(_this, _ret);
-  }
+  var builderEvents = [{ name: 'saveComponent', action: emit('onSaveComponent') }, { name: 'updateComponent', action: emit('onUpdateComponent') }, { name: 'removeComponent', action: emit('onDeleteComponent') }, { name: 'cancelComponent', action: emit('onUpdateComponent') }, { name: 'editComponent', action: emit('onEditComponent') }, { name: 'addComponent', action: onChange }, { name: 'saveComponent', action: onChange }, { name: 'updateComponent', action: onChange }, { name: 'removeComponent', action: onChange }, { name: 'deleteComponent', action: onChange }, { name: 'pdfUploaded', action: onChange }];
 
-  return FormBuilder;
-}(_react.Component);
+  var initializeBuilder = function initializeBuilder(builderProps) {
+    var options = builderProps.options,
+        form = builderProps.form;
+    var Builder = builderProps.Builder;
+
+    options = Object.assign({}, options);
+    form = Object.assign({}, form);
+
+    builder = new Builder(element.firstChild, form, options);
+    builderReady = builder.ready;
+
+    builderReady.then(function () {
+      onChange();
+      builderEvents.forEach(function (_ref) {
+        var name = _ref.name,
+            action = _ref.action;
+        return builder.instance.on(name, action);
+      });
+    });
+  };
+
+  (0, _react.useEffect)(function () {
+    initializeBuilder(props);
+    return function () {
+      return builder ? builder.instance.destroy(true) : null;
+    };
+  }, [props.form.display, props.form.components, props.options]);
+
+  return _react2.default.createElement(
+    'div',
+    { ref: function ref(el) {
+        return element = el;
+      } },
+    _react2.default.createElement('div', null)
+  );
+};
 
 FormBuilder.defaultProps = {
   options: {},
   Builder: _FormBuilder2.default
 };
+
 FormBuilder.propTypes = {
   form: _propTypes2.default.object,
   options: _propTypes2.default.object,
@@ -67,84 +102,6 @@ FormBuilder.propTypes = {
   onCancelComponent: _propTypes2.default.func,
   onEditComponent: _propTypes2.default.func,
   Builder: _propTypes2.default.any
-};
-
-var _initialiseProps = function _initialiseProps() {
-  var _this2 = this;
-
-  this.componentDidMount = function () {
-    _this2.initializeBuilder(_this2.props);
-  };
-
-  this.componentWillUnmount = function () {
-    if (_this2.builder !== undefined) {
-      _this2.builder.instance.destroy(true);
-    }
-  };
-
-  this.initializeBuilder = function (props) {
-    var options = Object.assign({}, props.options);
-    var form = Object.assign({}, props.form);
-    var Builder = props.Builder;
-
-    if (_this2.builder !== undefined) {
-      _this2.builder.instance.destroy(true);
-    }
-
-    _this2.builder = new Builder(_this2.element.firstChild, form, options);
-    _this2.builderReady = _this2.builder.ready;
-
-    _this2.builderReady.then(function () {
-      _this2.builder.instance.on('saveComponent', _this2.emit('onSaveComponent'));
-      _this2.builder.instance.on('updateComponent', _this2.emit('onUpdateComponent'));
-      _this2.builder.instance.on('removeComponent', _this2.emit('onDeleteComponent'));
-      _this2.builder.instance.on('cancelComponent', _this2.emit('onCancelComponent'));
-      _this2.builder.instance.on('editComponent', _this2.emit('onEditComponent'));
-      _this2.builder.instance.on('addComponent', _this2.onChange);
-      _this2.builder.instance.on('saveComponent', _this2.onChange);
-      _this2.builder.instance.on('updateComponent', _this2.onChange);
-      _this2.builder.instance.on('removeComponent', _this2.onChange);
-      _this2.builder.instance.on('deleteComponent', _this2.onChange);
-      _this2.builder.instance.on('pdfUploaded', _this2.onChange);
-    });
-  };
-
-  this.componentWillReceiveProps = function (nextProps) {
-    var _props = _this2.props,
-        options = _props.options,
-        form = _props.form;
-
-
-    if (form.display !== nextProps.form.display || options !== nextProps.options || form.components !== nextProps.form.components) {
-      _this2.initializeBuilder(nextProps);
-    }
-  };
-
-  this.render = function () {
-    return _react2.default.createElement(
-      'div',
-      { ref: function ref(element) {
-          return _this2.element = element;
-        } },
-      _react2.default.createElement('div', null)
-    );
-  };
-
-  this.onChange = function () {
-    if (_this2.props.hasOwnProperty('onChange') && typeof _this2.props.onChange === 'function') {
-      _this2.props.onChange(_this2.builder.instance.form);
-    }
-  };
-
-  this.emit = function (funcName) {
-    return function () {
-      if (_this2.props.hasOwnProperty(funcName) && typeof _this2.props[funcName] === 'function') {
-        var _props2;
-
-        (_props2 = _this2.props)[funcName].apply(_props2, arguments);
-      }
-    };
-  };
 };
 
 exports.default = FormBuilder;
