@@ -97,6 +97,187 @@ function Grid(props) {
 
   var normalizedPageSizes = pageSizes.map(normalizePageSize);
 
+  var getColumn = function getColumn(column) {
+    var key = column.key,
+        _column$sort = column.sort,
+        sort = _column$sort === undefined ? false : _column$sort,
+        _column$title = column.title,
+        title = _column$title === undefined ? '' : _column$title,
+        width = column.width;
+
+    var className = 'col col-md-' + width;
+
+    var columnProps = {
+      key: key,
+      className: className
+    };
+
+    if (!title) {
+      return _react2.default.createElement('div', columnProps);
+    }
+
+    if (!sort) {
+      return _react2.default.createElement(
+        'div',
+        columnProps,
+        _react2.default.createElement(
+          'strong',
+          null,
+          title
+        )
+      );
+    }
+
+    var sortKey = (0, _isString3.default)(sort) ? sort : key;
+    var ascSort = sortKey;
+    var descSort = '-' + sortKey;
+
+    var sortClass = '';
+    if (sortOrder === ascSort) {
+      sortClass = 'glyphicon glyphicon-triangle-top fa fa-caret-up';
+    } else if (sortOrder === descSort) {
+      sortClass = 'glyphicon glyphicon-triangle-bottom fa fa-caret-down';
+    }
+
+    return _react2.default.createElement(
+      'div',
+      columnProps,
+      _react2.default.createElement(
+        'span',
+        {
+          style: { cursor: 'pointer' },
+          onClick: function onClick() {
+            return onSort(column);
+          }
+        },
+        _react2.default.createElement(
+          'strong',
+          null,
+          title,
+          ' ',
+          _react2.default.createElement('span', { className: sortClass })
+        )
+      )
+    );
+  };
+
+  var getItem = function getItem(item) {
+    return _react2.default.createElement(
+      'li',
+      { className: 'list-group-item', key: item._id },
+      _react2.default.createElement(
+        'div',
+        { className: 'row', onClick: function onClick() {
+            return onAction(item, 'row');
+          } },
+        columns.map(function (column) {
+          return _react2.default.createElement(
+            'div',
+            { key: column.key, className: 'col col-md-' + column.width },
+            _react2.default.createElement(Cell, { row: item, column: column })
+          );
+        })
+      )
+    );
+  };
+
+  var PageSizeSelector = function PageSizeSelector() {
+    return _react2.default.createElement(
+      'div',
+      { className: 'col-auto' },
+      _react2.default.createElement(
+        'div',
+        { className: 'row align-items-center' },
+        _react2.default.createElement(
+          'div',
+          { className: 'col-auto' },
+          _react2.default.createElement(
+            'select',
+            {
+              className: 'form-control',
+              value: pageSize,
+              onChange: function onChange(event) {
+                return onPageSizeChanged(event.target.value);
+              }
+            },
+            normalizedPageSizes.map(function (_ref4) {
+              var label = _ref4.label,
+                  value = _ref4.value;
+              return _react2.default.createElement(
+                'option',
+                { key: value, value: value },
+                label
+              );
+            })
+          )
+        ),
+        _react2.default.createElement(
+          'span',
+          { className: 'col-auto' },
+          'items per page'
+        )
+      )
+    );
+  };
+
+  var FooterPagination = function FooterPagination() {
+    return _react2.default.createElement(
+      'div',
+      { className: 'col-auto' },
+      _react2.default.createElement(
+        'div',
+        { className: 'row align-items-center' },
+        _react2.default.createElement(
+          'div',
+          { className: 'col-auto' },
+          _react2.default.createElement(_Pagination2.default, {
+            pages: pages,
+            activePage: activePage,
+            pageNeighbours: pageNeighbours,
+            prev: 'Previous',
+            next: 'Next',
+            onSelect: onPage
+          })
+        ),
+        renderPageSizeSelector(props) ? _react2.default.createElement(PageSizeSelector, null) : null
+      )
+    );
+  };
+
+  var ItemCounter = function ItemCounter() {
+    return _react2.default.createElement(
+      'div',
+      { className: 'col-auto ml-auto' },
+      _react2.default.createElement(
+        'span',
+        { className: 'item-counter pull-right' },
+        _react2.default.createElement(
+          'span',
+          { className: 'page-num' },
+          firstItem,
+          ' - ',
+          lastItem
+        ),
+        ' / ',
+        total,
+        ' total'
+      )
+    );
+  };
+
+  var Footer = function Footer() {
+    return _react2.default.createElement(
+      'li',
+      { className: 'list-group-item' },
+      _react2.default.createElement(
+        'div',
+        { className: 'row align-items-center' },
+        renderPagination(props) ? _react2.default.createElement(FooterPagination, null) : null,
+        renderItemCounter(props) ? _react2.default.createElement(ItemCounter, null) : null
+      )
+    );
+  };
+
   return _react2.default.createElement(
     'div',
     null,
@@ -109,172 +290,11 @@ function Grid(props) {
         _react2.default.createElement(
           'div',
           { className: 'row' },
-          columns.map(function (column) {
-            var key = column.key,
-                _column$sort = column.sort,
-                sort = _column$sort === undefined ? false : _column$sort,
-                _column$title = column.title,
-                title = _column$title === undefined ? '' : _column$title,
-                width = column.width;
-
-            var className = 'col col-md-' + width;
-
-            var columnProps = {
-              key: key,
-              className: className
-            };
-
-            if (!title) {
-              return _react2.default.createElement('div', columnProps);
-            }
-
-            if (!sort) {
-              return _react2.default.createElement(
-                'div',
-                columnProps,
-                _react2.default.createElement(
-                  'strong',
-                  null,
-                  title
-                )
-              );
-            }
-
-            var sortKey = (0, _isString3.default)(sort) ? sort : key;
-            var ascSort = sortKey;
-            var descSort = '-' + sortKey;
-
-            var sortClass = '';
-            if (sortOrder === ascSort) {
-              sortClass = 'glyphicon glyphicon-triangle-top fa fa-caret-up';
-            } else if (sortOrder === descSort) {
-              sortClass = 'glyphicon glyphicon-triangle-bottom fa fa-caret-down';
-            }
-
-            return _react2.default.createElement(
-              'div',
-              columnProps,
-              _react2.default.createElement(
-                'span',
-                {
-                  style: { cursor: 'pointer' },
-                  onClick: function onClick() {
-                    return onSort(column);
-                  }
-                },
-                _react2.default.createElement(
-                  'strong',
-                  null,
-                  title,
-                  ' ',
-                  _react2.default.createElement('span', { className: sortClass })
-                )
-              )
-            );
-          })
+          columns.map(getColumn)
         )
       ),
-      items.map(function (item) {
-        return _react2.default.createElement(
-          'li',
-          { className: 'list-group-item', key: item._id },
-          _react2.default.createElement(
-            'div',
-            { className: 'row', onClick: function onClick() {
-                return onAction(item, 'row');
-              } },
-            columns.map(function (column) {
-              return _react2.default.createElement(
-                'div',
-                { key: column.key, className: 'col col-md-' + column.width },
-                _react2.default.createElement(Cell, { row: item, column: column })
-              );
-            })
-          )
-        );
-      }),
-      renderFooter(props) ? _react2.default.createElement(
-        'li',
-        { className: 'list-group-item' },
-        _react2.default.createElement(
-          'div',
-          { className: 'row align-items-center' },
-          renderPagination(props) ? _react2.default.createElement(
-            'div',
-            { className: 'col-auto' },
-            _react2.default.createElement(
-              'div',
-              { className: 'row align-items-center' },
-              _react2.default.createElement(
-                'div',
-                { className: 'col-auto' },
-                _react2.default.createElement(_Pagination2.default, {
-                  pages: pages,
-                  activePage: activePage,
-                  pageNeighbours: pageNeighbours,
-                  prev: 'Previous',
-                  next: 'Next',
-                  onSelect: onPage
-                })
-              ),
-              renderPageSizeSelector(props) ? _react2.default.createElement(
-                'div',
-                { className: 'col-auto' },
-                _react2.default.createElement(
-                  'div',
-                  { className: 'row align-items-center' },
-                  _react2.default.createElement(
-                    'div',
-                    { className: 'col-auto' },
-                    _react2.default.createElement(
-                      'select',
-                      {
-                        className: 'form-control',
-                        value: pageSize,
-                        onChange: function onChange(event) {
-                          return onPageSizeChanged(event.target.value);
-                        }
-                      },
-                      normalizedPageSizes.map(function (_ref4) {
-                        var label = _ref4.label,
-                            value = _ref4.value;
-                        return _react2.default.createElement(
-                          'option',
-                          { key: value, value: value },
-                          label
-                        );
-                      })
-                    )
-                  ),
-                  _react2.default.createElement(
-                    'span',
-                    { className: 'col-auto' },
-                    'items per page'
-                  )
-                )
-              ) : null
-            )
-          ) : null,
-          renderItemCounter(props) ? _react2.default.createElement(
-            'div',
-            { className: 'col-auto ml-auto' },
-            _react2.default.createElement(
-              'span',
-              { className: 'item-counter pull-right' },
-              _react2.default.createElement(
-                'span',
-                { className: 'page-num' },
-                firstItem,
-                ' - ',
-                lastItem
-              ),
-              ' / ',
-              total,
-              ' total'
-            )
-          ) : null
-        )
-      ) : null
+      items.map(getItem),
+      renderFooter(props) ? _react2.default.createElement(Footer, null) : null
     ) : _react2.default.createElement(
       'div',
       null,
