@@ -94,15 +94,14 @@ function Pagination(_ref2) {
         'li',
         { className: 'page-item ' + (activePage === 1 ? 'disabled' : '') },
         _react2.default.createElement(
-          'a',
+          'button',
           {
             className: 'page-link',
             onClick: function onClick() {
               if (activePage !== 1) {
                 onSelect(activePage - 1);
               }
-            },
-            href: 'javascript:void(0)'
+            }
           },
           prev
         )
@@ -130,13 +129,12 @@ function Pagination(_ref2) {
           'li',
           { className: 'page-item ' + className, key: page },
           _react2.default.createElement(
-            'a',
+            'button',
             {
               className: 'page-link',
               onClick: function onClick() {
                 return onSelect(page);
-              },
-              href: 'javascript:void(0)'
+              }
             },
             page
           )
@@ -146,15 +144,14 @@ function Pagination(_ref2) {
         'li',
         { className: 'page-item ' + (activePage === pages ? 'disabled' : '') },
         _react2.default.createElement(
-          'a',
+          'button',
           {
             className: 'page-link',
             onClick: function onClick() {
               if (activePage !== pages) {
                 onSelect(activePage + 1);
               }
-            },
-            href: 'javascript:void(0)'
+            }
           },
           next
         )
