@@ -209,7 +209,7 @@ var ReactComponent = function (_Field) {
     value: function checkValidity(data, dirty, rowData) {
       var valid = _get(ReactComponent.prototype.__proto__ || Object.getPrototypeOf(ReactComponent.prototype), 'checkValidity', this).call(this, data, dirty, rowData);
       if (!valid) {
-        return !valid;
+        return false;
       }
       return this.validate(data, dirty, rowData);
     }
