@@ -44,175 +44,179 @@ var _Grid2 = _interopRequireDefault(_Grid);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+var SubmissionGrid = function SubmissionGrid(props) {
+  var getSortQuery = function getSortQuery(key, sort) {
+    var currentSort = props.submissions.sort;
 
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+    var sortKey = (0, _isString3.default)(sort) ? sort : key;
+    var ascSort = sortKey;
+    var descSort = '-' + sortKey;
+    var noSort = '';
 
-var SubmissionGrid = function (_React$Component) {
-  _inherits(SubmissionGrid, _React$Component);
+    if (currentSort === ascSort) {
+      return descSort;
+    } else if (currentSort === descSort) {
+      return noSort;
+    } else {
+      return ascSort;
+    }
+  };
 
-  function SubmissionGrid() {
-    var _ref;
+  var onSort = function onSort(_ref) {
+    var key = _ref.key,
+        sort = _ref.sort;
 
-    var _temp, _this, _ret;
-
-    _classCallCheck(this, SubmissionGrid);
-
-    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
-      args[_key] = arguments[_key];
+    if ((0, _isFunction3.default)(sort)) {
+      return sort();
     }
 
-    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_ref = SubmissionGrid.__proto__ || Object.getPrototypeOf(SubmissionGrid)).call.apply(_ref, [this].concat(args))), _this), _this.onSort = function (_ref2) {
-      var key = _ref2.key,
-          sort = _ref2.sort;
+    var getSubmissions = props.getSubmissions;
 
-      if ((0, _isFunction3.default)(sort)) {
-        return sort();
+
+    getSubmissions(1, {
+      sort: getSortQuery(key, sort)
+    });
+  };
+
+  var getColumns = function getColumns(form) {
+    var columns = [];
+
+    _utils2.default.eachComponent(form.components, function (component) {
+      if (component.input && component.tableView && component.key) {
+        columns.push((0, _utils3.getComponentDefaultColumn)(component));
       }
+    });
 
-      var _this$props = _this.props,
-          getSubmissions = _this$props.getSubmissions,
-          currentSort = _this$props.submissions.sort;
+    columns.push({
+      key: 'operations',
+      title: 'Operations'
+    });
+
+    (0, _utils3.setColumnsWidth)(columns);
+
+    return columns;
+  };
+
+  var Icon = function Icon(_ref2) {
+    var icon = _ref2.icon;
+    return _react2.default.createElement(
+      'span',
+      null,
+      _react2.default.createElement('i', { className: 'fa fa-' + icon }),
+      '\xA0'
+    );
+  };
+
+  var OperationButton = function OperationButton(_ref3) {
+    var action = _ref3.action,
+        buttonType = _ref3.buttonType,
+        icon = _ref3.icon,
+        title = _ref3.title,
+        onAction = _ref3.onAction,
+        submission = _ref3.submission;
+    return _react2.default.createElement(
+      'span',
+      {
+        className: 'btn btn-' + buttonType + ' btn-sm form-btn',
+        onClick: (0, _utils3.stopPropagationWrapper)(function () {
+          return onAction(submission, action);
+        })
+      },
+      icon ? _react2.default.createElement(Icon, { icon: icon }) : null,
+      title
+    );
+  };
+
+  var Cell = function Cell(_ref4) {
+    var submission = _ref4.row,
+        column = _ref4.column;
+    var form = props.form,
+        onAction = props.onAction,
+        operations = props.operations;
 
 
-      var sortKey = (0, _isString3.default)(sort) ? sort : key;
-      var ascSort = sortKey;
-      var descSort = '-' + sortKey;
-      var noSort = '';
-
-      var nextSort = noSort;
-      if (currentSort === ascSort) {
-        nextSort = descSort;
-      } else if (currentSort === descSort) {
-        nextSort = noSort;
-      } else {
-        nextSort = ascSort;
-      }
-
-      getSubmissions(1, {
-        sort: nextSort
-      });
-    }, _this.getColumns = function (form) {
-      var columns = [];
-
-      _utils2.default.eachComponent(form.components, function (component) {
-        if (component.input && component.tableView && component.key) {
-          columns.push((0, _utils3.getComponentDefaultColumn)(component));
-        }
-      });
-
-      columns.push({
-        key: 'operations',
-        title: 'Operations'
-      });
-
-      (0, _utils3.setColumnsWidth)(columns);
-
-      return columns;
-    }, _this.Cell = function (_ref3) {
-      var submission = _ref3.row,
-          column = _ref3.column;
-      var _this$props2 = _this.props,
-          form = _this$props2.form,
-          onAction = _this$props2.onAction,
-          operations = _this$props2.operations;
-
-
-      if (column.key === 'operations') {
-        return _react2.default.createElement(
-          'div',
-          null,
-          operations.map(function (_ref4) {
-            var action = _ref4.action,
-                _ref4$buttonType = _ref4.buttonType,
-                buttonType = _ref4$buttonType === undefined ? 'primary' : _ref4$buttonType,
-                _ref4$icon = _ref4.icon,
-                icon = _ref4$icon === undefined ? '' : _ref4$icon,
-                _ref4$permissionsReso = _ref4.permissionsResolver,
-                permissionsResolver = _ref4$permissionsReso === undefined ? function () {
-              return true;
-            } : _ref4$permissionsReso,
-                _ref4$title = _ref4.title,
-                title = _ref4$title === undefined ? '' : _ref4$title;
-            return permissionsResolver(form, submission) ? _react2.default.createElement(
-              'span',
-              {
-                className: 'btn btn-' + buttonType + ' btn-sm form-btn',
-                onClick: (0, _utils3.stopPropagationWrapper)(function () {
-                  return onAction(submission, action);
-                }),
-                key: action
-              },
-              icon ? _react2.default.createElement(
-                'span',
-                null,
-                _react2.default.createElement('i', { className: 'fa fa-' + icon }),
-                '\xA0'
-              ) : null,
-              title
-            ) : null;
-          })
-        );
-      }
-
-      var value = (0, _isFunction3.default)(column.value) ? column.value(submission) : (0, _get3.default)(submission, column.key, '');
-
-      return (0, _isObject3.default)(value) && value.content ? value.isHtml ? _react2.default.createElement('div', { dangerouslySetInnerHTML: { __html: value.content } }) : _react2.default.createElement(
-        'span',
+    if (column.key === 'operations') {
+      return _react2.default.createElement(
+        'div',
         null,
-        String(value.content)
-      ) : _react2.default.createElement(
-        'span',
-        null,
-        String(value)
+        operations.map(function (_ref5) {
+          var action = _ref5.action,
+              _ref5$buttonType = _ref5.buttonType,
+              buttonType = _ref5$buttonType === undefined ? 'primary' : _ref5$buttonType,
+              _ref5$icon = _ref5.icon,
+              icon = _ref5$icon === undefined ? '' : _ref5$icon,
+              _ref5$permissionsReso = _ref5.permissionsResolver,
+              permissionsResolver = _ref5$permissionsReso === undefined ? function () {
+            return true;
+          } : _ref5$permissionsReso,
+              _ref5$title = _ref5.title,
+              title = _ref5$title === undefined ? '' : _ref5$title;
+          return permissionsResolver(form, submission) ? _react2.default.createElement(OperationButton, {
+            key: action,
+            action: action,
+            buttonType: buttonType,
+            icon: icon,
+            title: title,
+            submission: submission,
+            onAction: onAction
+          }) : null;
+        })
       );
-    }, _this.render = function () {
-      var _this$props3 = _this.props,
-          propColumns = _this$props3.columns,
-          form = _this$props3.form,
-          getSubmissions = _this$props3.getSubmissions,
-          onAction = _this$props3.onAction,
-          onPageSizeChanged = _this$props3.onPageSizeChanged,
-          pageSizes = _this$props3.pageSizes,
-          _this$props3$submissi = _this$props3.submissions,
-          limit = _this$props3$submissi.limit,
-          _this$props3$submissi2 = _this$props3$submissi.pagination,
-          page = _this$props3$submissi2.page,
-          numPages = _this$props3$submissi2.numPages,
-          total = _this$props3$submissi2.total,
-          sort = _this$props3$submissi.sort,
-          submissions = _this$props3$submissi.submissions;
+    }
+
+    var value = (0, _isFunction3.default)(column.value) ? column.value(submission) : (0, _get3.default)(submission, column.key, '');
+
+    return (0, _isObject3.default)(value) && value.content ? value.isHtml ? _react2.default.createElement('div', { dangerouslySetInnerHTML: { __html: value.content } }) : _react2.default.createElement(
+      'span',
+      null,
+      String(value.content)
+    ) : _react2.default.createElement(
+      'span',
+      null,
+      String(value)
+    );
+  };
+
+  var propColumns = props.columns,
+      form = props.form,
+      getSubmissions = props.getSubmissions,
+      onAction = props.onAction,
+      onPageSizeChanged = props.onPageSizeChanged,
+      pageSizes = props.pageSizes,
+      _props$submissions = props.submissions,
+      limit = _props$submissions.limit,
+      _props$submissions$pa = _props$submissions.pagination,
+      page = _props$submissions$pa.page,
+      numPages = _props$submissions$pa.numPages,
+      total = _props$submissions$pa.total,
+      sort = _props$submissions.sort,
+      submissions = _props$submissions.submissions;
 
 
-      var columns = propColumns.length ? propColumns : _this.getColumns(form);
-      var skip = (page - 1) * limit;
-      var last = Math.min(skip + limit, total);
+  var columns = propColumns.length ? propColumns : getColumns(form);
+  var skip = (page - 1) * limit;
+  var last = Math.min(skip + limit, total);
 
-      return _react2.default.createElement(_Grid2.default, {
-        Cell: _this.Cell,
-        activePage: page,
-        columns: columns,
-        emptyText: 'No data found',
-        firstItem: skip + 1,
-        items: submissions,
-        lastItem: last,
-        onAction: onAction,
-        onPage: getSubmissions,
-        onPageSizeChanged: onPageSizeChanged,
-        onSort: _this.onSort,
-        pageSize: limit,
-        pageSizes: pageSizes,
-        pages: numPages,
-        sortOrder: sort,
-        total: total
-      });
-    }, _temp), _possibleConstructorReturn(_this, _ret);
-  }
-
-  return SubmissionGrid;
-}(_react2.default.Component);
+  return _react2.default.createElement(_Grid2.default, {
+    Cell: Cell,
+    activePage: page,
+    columns: columns,
+    emptyText: 'No data found',
+    firstItem: skip + 1,
+    items: submissions,
+    lastItem: last,
+    onAction: onAction,
+    onPage: getSubmissions,
+    onPageSizeChanged: onPageSizeChanged,
+    onSort: onSort,
+    pageSize: limit,
+    pageSizes: pageSizes,
+    pages: numPages,
+    sortOrder: sort,
+    total: total
+  });
+};
 
 SubmissionGrid.propTypes = {
   columns: _types.Columns,
@@ -224,6 +228,7 @@ SubmissionGrid.propTypes = {
   pageSizes: _types.PageSizes,
   submissions: _propTypes2.default.object.isRequired
 };
+
 SubmissionGrid.defaultProps = {
   columns: [],
   getSubmissions: function getSubmissions() {},
@@ -257,4 +262,5 @@ SubmissionGrid.defaultProps = {
   }],
   pageSizes: _constants.defaultPageSizes
 };
+
 exports.default = SubmissionGrid;
