@@ -65,7 +65,7 @@ var formAccessUser = function formAccessUser(formAccess) {
 
 var projectAccessUser = function projectAccessUser(projectAccess) {
   return {
-    type: type.USER_FORM_ACCESS,
+    type: type.USER_PROJECT_ACCESS,
     projectAccess: projectAccess
   };
 };
